+++ internal/js_ast/js_ast.go
@@ -1021,17 +1021,6 @@ type SContinue struct {
 	Label *LocRef
 }
 
-func IsSuperCall(stmt Stmt) bool {
-	if expr, ok := stmt.Data.(*SExpr); ok {
-		if call, ok := expr.Value.Data.(*ECall); ok {
-			if _, ok := call.Target.Data.(*ESuper); ok {
-				return true
-			}
-		}
-	}
-	return false
-}
-
 type ClauseItem struct {
 	Alias string